下载代码整合; (多线程下载速度慢,目前原因未知);

+++ JMCLCode/src/main/java/utils/Utils.java
@@ -92,7 +92,7 @@ public class Utils {
         return null;
     }
 
-    public static byte[] getPartOfTheFileContent(URL url, int start, int end) throws IOException {
+    public static byte[] getPartOfTheFileContent(URL url, int start, long end) throws IOException {
         URLConnection connection = url.openConnection();
         if (connection instanceof HttpURLConnection httpConnection) {
             httpConnection.setRequestMethod("GET");
